修复 ProactChannel::handle_connect_completed() 触发 unregister_handler() 从而导致 handler->_write_queue 中的 IORequest 被提前释放，从而崩溃的问题

+++ src/test_loofah/manually/test_proact_package_manually.cpp
@@ -184,7 +184,7 @@ void test_proact_package_manually()
     while (!prepared || server != nullptr || client != nullptr)
     {
         const uint64_t idle_ms = std::min<uint64_t>(
-            60 * 1000, std::max<uint64_t>(TimeWheel::RESOLUTION_MS, timewheel.get_idle()));
+            60 * 1000, std::max<uint64_t>(unsigned(TimeWheel::RESOLUTION_MS), timewheel.get_idle()));
         if (proactor.handle_events(idle_ms) < 0)
             break;
         timewheel.tick();
