+++ src/test_loofah/manually/test_react_package_manually.cpp
@@ -192,7 +192,7 @@ void test_react_package_manually()
     while (!prepared || server != nullptr || client != nullptr)
     {
         const uint64_t idle_ms = std::min<uint64_t>(
-            60 * 1000, std::max<uint64_t>(TimeWheel::RESOLUTION_MS, timewheel.get_idle()));
+            60 * 1000, std::max<uint64_t>(unsigned(TimeWheel::RESOLUTION_MS), timewheel.get_idle()));
         if (reactor.handle_events(idle_ms) < 0)
             break;
         timewheel.tick();
