+++ src/test_loofah/rst/test_proact_package_rst.cpp
@@ -203,7 +203,7 @@ class TestProactPackageRST : public TestFixture
         while (!prepared || server != nullptr || client != nullptr)
         {
             const uint64_t idle_ms = std::min<uint64_t>(
-                60 * 1000, std::max<uint64_t>(TimeWheel::RESOLUTION_MS, timewheel.get_idle()));
+                60 * 1000, std::max<uint64_t>(unsigned(TimeWheel::RESOLUTION_MS), timewheel.get_idle()));
             if (proactor.handle_events(idle_ms) < 0)
                 break;
             timewheel.tick();
